Cleaned up stuff + fixed warnings

+++ firmware/eeprommer3/input.cpp
@@ -6,7 +6,7 @@
 #include "input.hpp"
 #include "tft.hpp"
 
-TouchCtrl::TouchCtrl(uint8_t xp, uint8_t xm, uint8_t yp, uint8_t ym, uint8_t resist)
+TouchCtrl::TouchCtrl(uint16_t xp, uint16_t xm, uint16_t yp, uint16_t ym, uint16_t resist)
   : TouchScreen(xp, yp, xm, ym, resist) {
   // Empty
 }
